// src/edgeiq-os/services/performance-intelligence/types.ts
export type PerformanceIntelligenceManifest = {
  pass_token?: string;
  generated_timestamp?: string;
  feed_version?: string;
  row_counts?: Record<string, number>;
  unsupported_fields?: string[];
  known_limitations?: string[];
  source_snapshots?: string[];
};

export type RaceIntelligenceIndexRow = {
  meeting_key: string;
  meeting: string;
  race_context_key: string;
  race_date: string;
  state: string;
  track: string;
  race_number: string;
  race_name: string;
  distance_metres: string;
  race_class: string;
  track_condition: string;
  track_rating: string;
  rail_position: string;
  runner_count: string;
  race_quality_state: string;
  governed_race_time_seconds: string;
  selected_benchmark_id: string;
  selected_benchmark_level: string;
  selected_benchmark_confidence: string;
  selected_benchmark_sample_size: string;
  benchmark_time_seconds: string;
  seconds_vs_benchmark: string;
  fallback_path: string;
  fingerprint_pattern: string;
  fingerprint_quality_state: string;
  source_snapshot_ids?: string[];
  generated_timestamp: string;
};

export type HorseIntelligenceIndexRow = {
  race_context_key: string;
  runner_number: string;
  horse_identity_evidence_id: string;
  horse_code: string;
  horse_name: string;
  identity_quality_state: string;
  profile_quality_state: string;
  performance_count: string;
  eligible_performance_count: string;
  date_range: string;
  track_profile: string;
  distance_profile: string;
  going_profile: string;
  class_profile: string;
  pressure_profile: string;
  acceleration_profile: string;
  late_speed_profile: string;
  efficiency_profile: string;
  consistency_profile: string;
  track_affinity: string;
  distance_affinity: string;
  pattern_distribution: string;
  quality_distribution: string;
  source_snapshot_ids?: string[];
  generated_timestamp: string;
};

export type HistoricalPerformanceIntelligenceRow = {
  current_race_context_key: string;
  runner_number: string;
  performance_fact_id: string;
  legacy_performance_fact_id: string;
  race_context_key: string;
  race_date: string;
  track: string;
  state: string;
  race_number: string;
  race_name: string;
  race_class: string;
  distance_metres: string;
  track_condition: string;
  track_rating: string;
  rail_position: string;
  horse: string;
  horse_code: string;
  trainer: string;
  jockey: string;
  barrier: string;
  weight: string;
  finish_position: string;
  margin_raw: string;
  margin_lengths_raw: string;
  starting_price: string;
  governed_race_time_seconds: string;
  performance_quality_state: string;
  benchmark_exclusion_reason: string;
  source_snapshot_ids?: string[];
  selected_benchmark_id: string;
  benchmark_level: string;
  benchmark_confidence: string;
  benchmark_sample_size: string;
  benchmark_time_seconds: string;
  race_seconds_vs_benchmark: string;
  fallback_path: string;
  fingerprint_pattern: string;
  fingerprint_quality_state: string;
};

export type BenchmarkExplanationIndexRow = {
  benchmark_id: string;
  benchmark_level: string;
  hierarchy_dimensions: string;
  track: string;
  course: string;
  distance_metres: string;
  race_class_canonical: string;
  going_canonical: string;
  rail_canonical: string;
  sample_size: string;
  eligible_sample_size: string;
  excluded_sample_size: string;
  mean_time_seconds: string;
  median_time_seconds: string;
  trimmed_mean_time_seconds: string;
  standard_deviation_seconds: string;
  median_absolute_deviation_seconds: string;
  lower_quantile_seconds: string;
  upper_quantile_seconds: string;
  confidence_state: string;
  confidence_score: string;
  minimum_sample_rule: string;
  benchmark_version: string;
  source_snapshot_id: string;
  generated_timestamp: string;
};

export type PerformanceIntelligenceFeed = {
  manifest: PerformanceIntelligenceManifest;
  race_intelligence_index: RaceIntelligenceIndexRow[];
  horse_intelligence_index: HorseIntelligenceIndexRow[];
  historical_performance_intelligence_index: HistoricalPerformanceIntelligenceRow[];
  benchmark_explanation_index: BenchmarkExplanationIndexRow[];
};